+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+
+## [Unreleased]
+### Fixed
+- A bug with the omitted key chain seeds if the `seed` property present
+
+### Changed
+- Now default signer can be provided on the wallet creation
+
 ## [1.14.0-rc.3] - 2020-09-04
 ## Added
 - Create, update, remove data reviewable requests
@@ -29,7 +37,7 @@ for further information about branching and tagging conventions.
 
 ## [1.13.1-rc.0] - 2020-06-02
 ### Changed
-- Subject size from 256 to 1000 
+- Subject size from 256 to 1000
 
 ## [1.13.0] - 2020-04-13
 
@@ -105,7 +113,7 @@ for further information about branching and tagging conventions.
 - Integration tests
 
 ## [1.10.2-rc.0] - 2019-09-16
-### Added 
+### Added
 - Builder for `RemoveAssetOp`
 - Unit tests for `RemoveAssetOp`
 - Integration tests for `RemoveAssetOp`